fix(filter): Move from `data` to `filter`
Cherry-pick fe2fb1e606f81400cfb55a66515beee2d2f67602 (#288)
Cherry-pick 79e31673db8ce35c7bd0e59f95b732c983b329a6 (#285)
